feat/RSS-ECOMM-3_08: add useBreadcrumbs hook
