+++ src/models/breadcrumbs.model.ts
@@ -1,0 +1,5 @@
+export interface Breadcrumb {
+  label: string;
+  to: string;
+  isLast?: boolean;
+}
