+++ src/pages/Layout/components/Breadcrumbs/Breadcrumbs.component.tsx
@@ -1,17 +1,10 @@
 import { HideOnPathsComponent } from '@components/HideOnPaths/HideOnPaths.component';
-import useCategory from '@hooks/useCategory';
-import useProduct from '@hooks/useProduct';
+import { useBreadcrumbs } from '@hooks/useBreadcrumbs';
 import HomeIcon from '@mui/icons-material/Home';
 import { Breadcrumbs, Chip, ChipProps, CircularProgress, emphasize, styled, Toolbar, Typography } from '@mui/material';
 import { blue } from '@mui/material/colors';
 import { ElementType } from 'react';
-import { Link, LinkProps, useLocation } from 'react-router-dom';
-
-interface Breadcrumb {
-  label: string;
-  to: string;
-  isLast?: boolean;
-}
+import { Link, LinkProps } from 'react-router-dom';
 
 const hideBreadcrumbsPaths = ['/', '/login', '/registration', '/profile', '/about', '/cart'];
 const breadcrumbSpacing = 3;
@@ -41,36 +34,7 @@ const StyledBreadcrumb = styled(Chip)<ChipProps & { component: ElementType } & L
 });
 
 export default function BreadcrumbsComponent() {
-  const location = useLocation();
-  const { category } = useCategory();
-  const { product, isProductLoading } = useProduct();
-  const paths = location.pathname.split('/');
-
-  const breadcrumbs: Breadcrumb[] = [];
-
-  for (let i = 0; i < paths.length; i += 1) {
-    const breadcrumb: Breadcrumb = {
-      label: paths[i],
-      to: '/',
-    };
-
-    if (paths[i] === 'categories') {
-      breadcrumb.label = category?.name.en || '...';
-      i += 1;
-    }
-
-    if (paths[i] === 'products') {
-      breadcrumb.label = isProductLoading || !product?.name.en ? '...' : product.name.en;
-      i += 1;
-    }
-
-    breadcrumb.to = paths.slice(0, i + 1).join('/');
-    breadcrumb.isLast = i === paths.length - 1;
-
-    if (breadcrumb.label !== '') {
-      breadcrumbs.push(breadcrumb);
-    }
-  }
+  const { breadcrumbs } = useBreadcrumbs();
 
   return (
     <HideOnPathsComponent paths={hideBreadcrumbsPaths}>
